requirements.txt
# 基础依赖
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
aiohttp>=3.8.0
asyncio>=3.4.3

# 网页抓取和自动化
selenium>=4.8.0
webdriver-manager>=3.8.0

# 数据处理和解析
pandas>=1.5.0
bibtexparser>=1.4.0
PyYAML>=6.0

# PDF处理
PyPDF2>=3.0.0
pdfplumber>=0.7.0

# 日志和配置
loguru>=0.6.0
pydantic>=1.10.0
python-dotenv>=0.21.0

# 网络工具
urllib3>=1.26.0
fake-useragent>=1.2.0

# 异步处理
asyncio-throttle>=1.0.0
aiofiles>=22.0.0

# 命令行界面
click>=8.1.0
tqdm>=4.64.0
colorama>=0.4.0

# XML解析（用于arXiv、PubMed等）
xmltodict>=0.13.0

# 科学计算（可选，用于高级分析）
numpy>=1.24.0
scipy>=1.10.0

# 数据可视化（可选）
matplotlib>=3.6.0
seaborn>=0.12.0

# 自然语言处理（可选）
nltk>=3.8.0
spacy>=3.5.0

# 开发工具
pytest>=7.2.0
pytest-asyncio>=0.21.0
black>=22.0.0
flake8>=5.0.0